app/requirements.txt
Django==4.2.3
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
pillow